fix(orm): refine some logic

+++ pkg/orm/model/logic_epoch.go
@@ -3,5 +3,5 @@ package model
 // using transaction to generate increasing epoch
 type LogicEpoch struct {
 	Model
-	Epoch int `gorm:"type:bigint not null default 1"`
+	Epoch int64 `gorm:"type:bigint not null default 1"`
 }
